Add a make rule for using pip to create a bundle file for TiddlyWebWiki.
A bundle, in pip terms, is an archive of all the source packages
needed to install a particular thing.

So once you have a bundle you can pip install ./<bundlefile>
and your stuff gets installed. Without additional args the
packages in the bundle are installed to site-packages. In a
non-root situation it is possible to install the package to a
virtualenv with (assuming you are in the virtualenv):

   pip -E . /path/to/<bundlefile>

(note the indicator that this is path, not a name is required
otherwise pip will go looking for the package named bundlefile)