Importer and instancer are not really plugins, so stop treating them as they are.
This means that a bunch of functions now need access to the current tidldyweb.config.

These changes appears to accomplish that but there are some concerns:

* They probably change public APIs
* It's unclear that test coverage is sufficient to cover these, although
  anecdotal testing with building an instance shows things to be okay.

I did this while trying to clean up for including atom, noticing that
twanger_plugins in tiddlywebwiki.config was not actually being used in
any way.