Manage tiddlywebwiki.instance:store_contents in a more straightforward fashion
We know that store contents needs to include stuff from console.
We choose in this change to use stuff that is installed in the package,
(i.e. in the resources dir) rather than raw. This means that when
tiddlywebwiki is packaged up and distributed and an instance created the
instance has all the info it needs.