First stab and providing a configurable limit on sending binary tiddler content.
In the old days binary tiddlers were always links. Then we switched to
always base64. Now we can have a combination. Any binary tiddler which
is longer (in bytes) than config['tiddlywebwiki.binary_limit'] will be
sent as a link.

The diff is small and fairly self explanatory.

Right now the default is no limit.