Adjust wiki <title> html to text for performance.
Using html5lib is _slow_ for this particular instance, since
the amount and complexity of the text involved is low.

Further gains can be had by not doing the wiki -> html -> text
transformation on the candidate_title at all (because wikklytext
is non-optimal for small bits of text too), but instead just use
the wiki text.

Have not done that in this commit, though, because that could be
considered a functional change.

Please ask if the algorithm makes no sense.