Very rough spike at doing lazyloading by way of the serialization.
The basic idea here is that if a tiddler eligible to be lazy
it is not sent out with the initial wiki. Instead it's name is
added to a virtual tiddler called LazyTiddlers.

A plugin, LazyTiddlersPlugin, reads that tiddler when the wiki
boots and loads all the named tiddlers.

The purpose of this activity is to make the initially loaded wiki
smaller and to transform the single long request to load a wiki
into one medium request plus many small requests.

This has not yet been tested with a large number of tiddlers.

The definition of "eligible to be lazy" needs to be refined.