Small optimization so that the base tiddlywiki file is only read once.
Without this change, every request for a wiki reads the wiki file off
disk and then does its work. This keeps the contents of the file in
memory, avoiding the disk read.

On a personal tiddlywebwiki this is no big deal, but on tiddlyspace,
will probably save a few cycles here and there.