Get rid of wikification and then plain textification of HTML title of wiki.
It turns out this small piece of work, which is used so that the
download-ing wiki has a high fidelity title, is a _huge_ contributor
to the processing time of creating the wiki, for vanishingly small
value that is wiped out by tiddlywiki's own processing once it is
fully loaded.

So now it doesn't do that. Instead the raw wikitext sets the title.
If the wikitext has markup in it, such as //foobar// then the //
remain.

And the second highest cpu using leg of the code goes away!